+++ src/ready.lua
@@ -19,8 +19,21 @@ modutil.mod.Path.Wrap("BoonInfoPopulateTraits", function(base, screen)
 	end
 end)
 
+---Add or remove empty PlayerUnit (Melinoe) entry in the TraitDictionary when Codex is opened.<br>
+---Note: this entry controls the appearance of the Boon offering button.<br>
+---We start by checking if we're currently not inside a HubRoom, it is mandatory to check this<br>
+---because `game.CurrentRoom` keeps all its data from the previous run for some reason.<br>
+---Then we can safely check the gods encountered during the run. If any "slot" god appeared,<br>
+---the button can safely appear.
+---TODO: make the button appear if there are pinned boons.
+---@param base any
+modutil.mod.Path.Wrap("OpenCodexScreen", function(base)
+	OpenCodexScreen_UpdateMelinoeBoonOfferingButton()
+	base()
+end)
+
 -- Forces the Boon button to appear for Melinoe
-ScreenData.BoonInfo.TraitDictionary["PlayerUnit"] = {}
+
 
 --[[
 	Hard coded ordering for the Olympian boons, by type then by God